format and move sanitize-latex
