+++ src/components/content/entity.tsx
@@ -25,8 +25,8 @@ import { editorRenderers } from '@/serlo-editor/plugin/helpers/editor-renderer'
 import { isEmptyRowsDocument } from '@/serlo-editor/plugins/rows/utils/static-is-empty'
 import { CourseNavigation } from '@/serlo-editor/plugins/serlo-template-plugins/course/course-navigation'
 import { StaticRenderer } from '@/serlo-editor/static-renderer/static-renderer'
-import { createRenderers } from '@/serlo-editor-integration/create-renderers'
 import { isArticleDocument } from '@/serlo-editor/types/plugin-type-guards'
+import { createRenderers } from '@/serlo-editor-integration/create-renderers'
 
 export interface EntityProps {
   data: EntityData
