+++ src/serlo-editor-integration/serlo-plugin-wrappers/image-serlo-static-renderer.tsx
@@ -1,7 +1,7 @@
 import { useRouter } from 'next/router'
 
-import { ExtraInfoIfRevisionView } from '../extra-info-if-revision-view'
 import { EditorImageDocument } from '../../serlo-editor/types/editor-plugins'
+import { ExtraInfoIfRevisionView } from '../extra-info-if-revision-view'
 import { ImageStaticRenderer } from '@/serlo-editor/plugins/image/static'
 
 export function ImageSerloStaticRenderer(props: EditorImageDocument) {
