+++ src/serlo-editor/plugins/text/utils/sanitize-latex.ts
